ParkingLotValues.json har nu rätt setPath En UI skapades - testas än
