--- Core/ParkingLotLimitationValues.json
@@ -1,7 +1,0 @@
-{
-  "Per hour for Car": 100,
-  "Per hour for MC": 20,
-  "Parking spots in the garage": 25,
-  "Motorcycles in the same parking spot": 3,
-  "Cars in the same parking spot": 1
-}
