+++ PragueParking_2.0/ParkingGarageLimitationValues.cs
@@ -8,7 +8,6 @@ namespace PragueParking_2._0
         public int PerHourForCar { get; set; }
 
         [JsonPropertyName("Per hour for MC")]
-        
         public int PerHourForMC { get; set; }
